refactor: deduplicate serial console, operations check, and fix markers
- Extract shared serial console logic into aws/common/serial_console.py,
  slim down VM and BM stubs to thin wrappers
- Extract check_operations_passed() utility from CrudOperationsCheck,
  reuse in TopologyPlacementCheck
- Fix SerialConsoleCheck markers: ["vm"] → ["vm", "bm"]
